feat: Add 16-track support, modulators, microtiming, and UI enhancements
This commit adds extensive functionality to the PER|FORMER sequencer:

## 16-Track Support
- Extended MIDI routing from 8 to 16 tracks
- Added dual bank system with LED color inversion for bank 2
- Fixed SongPage and SystemPage to support all 16 tracks
- Extended MIDI output enums (GateSource, NoteSource, VelocitySource, ControlSource)
- Added enum validation to prevent project load crashes
- Note: Tracks limited to 8 patterns each (CONFIG_PATTERN_COUNT)
- Note: Song mode limited to 4 slots (CONFIG_SONG_SLOT_COUNT)

## 8 Modulators/LFOs
- Implemented 8 independent LFO modulators
- Shapes: Sine, Triangle, Saw Up/Down, Square, Smooth Random, Stepped Random
- Added ModulatorPage with waveform preview
- Routing popup for quick MIDI CC/Note mapping
- Preset: Mod 1-8 → MIDI 1-8 CC 0
- Custom preset: Tracks 9-16 → MIDI 1-8 @ velocity 127
- Triplet and dotted note rate divisions (1/64T, 1/32T, 1/16T, 1/8T, 1/4T, 1/32., 1/16., 1/8., 1/4., 1/2.)
- Smooth parameter (0-2000ms) for random shapes

## Microtiming Recording
- Added timingQuantize parameter (0-100%) to NoteTrack
- Captures gate timing offset (0-15) during live recording
- Quantize strength blends between natural feel and grid precision
- Stored in existing GateOffset field
- UI integration in Track settings page

## Performer Page Enhancements
- Pattern mode LED feedback: green=active, red=muted, yellow=selected
- Muted tracks show dimmed pattern number in pattern mode
- Fixed LED colors for both banks (no inversion for mute status)
- Improved all-tracks view visual feedback

## MIDI Output Improvements
- Display tracks 9-16, Mod 1-8, CV In 1-4 as sources
- CV inputs can now be mapped to MIDI CC outputs
- Fixed ControlSource display for modulators and CV inputs
- Enum clamping prevents crashes from invalid project data
- Context menu text shortened for readability

## UI/UX Polish
- Bank 2 LED inversion for sequence steps (red LEDs)
- Consistent LED behavior across all pages
- Improved context menus and navigation
- Better parameter display and editing

## Memory & Performance
- Firmware size: 335540 bytes text + 6196 data + 150948 bss
- Compatible with existing project format (ProjectVersion::Version27)
- Enum validation for backward compatibility

+++ src/apps/sequencer/model/Routing.h
@@ -546,7 +546,7 @@ public:
 
     // global state for keeping active set of routed targets
     static bool isRouted(Target target, int trackIndex = -1);
-    static void setRouted(Target target, uint8_t tracks, bool routed);
+    static void setRouted(Target target, uint16_t tracks, bool routed);
     static void printRouted(StringBuilder &str, Target target, int trackIndex = -1);
 
 private:
